requirements.txt
requests==2.31.0
gspread==6.0.2
google-auth==2.28.1
python-dotenv==1.0.1
